+++ test/cases/resolve-attr-json-require/src/main.js
@@ -1,0 +1,6 @@
+const elm = document.getElementById('test-elm');
+
+const value = elm.getAttribute('data-bigpicture');
+const data = JSON.parse(value);
+
+console.log('>> main', { value }, data);